+++ bashcmd/doc.go
@@ -1,2 +1,0 @@
-// Package bashcmd provides an Eino-compatible run_bash_command tool.
-package bashcmd