+++ fetchurl/doc.go
@@ -1,2 +1,0 @@
-// Package fetchurl provides an Eino-compatible fetch_url tool.
-package fetchurl